% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{collect.data,nowcast-method}
\alias{collect.data,nowcast-method}
\title{Collect data}
\usage{
\S4method{collect.data}{nowcast}(object, directory = NULL)
}
\arguments{
\item{nowcast}{}
}
\value{
nowcast
}
\description{
Collect data
}
