% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{collect.data.by.ticker}
\alias{collect.data.by.ticker}
\title{Collect by ticker}
\usage{
collect.data.by.ticker(ticker, directory)
}
\arguments{
\item{ticker}{character ticker}

\item{directory}{character path to database directory with subfolder "transform" (usually \code{Sys.getenv('directory')})}
}
\value{
data.table
}
\description{
Returns transformed data for specified ticker
}
\examples{
head(collect.data.by.ticker('gdp_real',
 Sys.getenv('directory')))
#           date gdp_real
# 1:  1995-01-01 10757.47
# 2:  1995-04-01 11072.24
# 3:  1995-07-01 11666.89
# 4:  1995-10-01 11845.80
# 5:  1996-01-01 10518.41
}
\keyword{internal}
