% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create.lag}
\alias{create.lag}
\title{Create lag}
\usage{
create.lag(df, lags)
}
\arguments{
\item{df}{data.frame}

\item{lags}{integer vector}
}
\value{
data.frame
}
\description{
Returns data.frame with lags of second column
}
\examples{
create.lag(data.frame("index"=c(1:5), x = c(1:5)), lags = c(1,2))
# index x x__lag1 x__lag2
#     3 3       2       1
#     4 4       3       2
#     5 5       4       3

}
\keyword{internal}
