% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{cut.df.from.start}
\alias{cut.df.from.start}
\title{Cut df}
\usage{
\method{cut}{df.from.start}(df, len)
}
\arguments{
\item{df}{data.frame}

\item{len}{integer}
}
\value{
data.frame
}
\description{
Removes first \code{nrow(df)}-\code{len} rows from df
}
\examples{
cut.df.from.start(data.frame(x=c(1:10)), len = 5)
#   x y
# 4 4 4
# 5 5 5
}
\keyword{internal}
