% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{export.nowcast.to.csv}
\alias{export.nowcast.to.csv}
\title{Export result to csv}
\usage{
export.nowcast.to.csv(input, output)
}
\arguments{
\item{input}{path to .Rda/.Rdata with nowcasting resuls}

\item{output}{path to .csv}
}
\description{
Export result to csv
}
