% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{filter.data.by.frequency}
\alias{filter.data.by.frequency}
\title{Filter by frequency}
\usage{
filter.data.by.frequency(df, object, predictor_freq)
}
\arguments{
\item{df}{data.frame without NA's}

\item{object}{nowcast}

\item{predictor_freq}{character from ('d','w', 'm', 'q')

If the target variable frequency and predictor frequency matches, returns original df.
If the target variable is quarterly and predictor is monthly, returns every third row
of df, ending with last df's row.
If the target variable is monthly and predictor is quarterly, repeats each df's row
three times.
If the predictor is daily or weekly and
the target variable is quarterly or monthly,
firstly for the df's last date defines day number in its quarter or months (\code{dn}),
then returns df, where each row corresponds to the \code{dn} day of quarter of month.}
}
\value{
data.frame
}
\description{
Returns a date frame with a frequency that
matches the target variable frequency (\code{object@freq})
}
\keyword{internal}
