% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{make.monthly}
\alias{make.monthly}
\title{Make monthly}
\usage{
make.monthly(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
data.frame
}
\description{
Transforms weekly or daily df to monthly df,
returning only last observation for each month
}
\keyword{internal}
