% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{remove.column}
\alias{remove.column}
\title{Remove columns}
\usage{
remove.column(df, column)
}
\arguments{
\item{df}{data.frame}

\item{column}{character column name}
}
\value{
data.frame
}
\description{
Removes specified \code{column} from df or breaks with message if \code{column} not in df's columns names
}
\examples{
remove.column(data.frame(x=c(1:3), y = c(1:3)), column = "y")
#   x
# 1 1
# 2 2
# 3 3
remove.column(data.frame(x=c(1:3), y = c(1:3)), column = "value")
#   Can't remove column
# value
# Error in remove.column(data.frame(x = c(1:3), y = c(1:3)), column = "value") :
# no loop for break/next, jumping to top level
}
\keyword{internal}
