% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{run.from.json}
\alias{run.from.json}
\title{Run from JSON}
\usage{
run.from.json(input, output = NULL)
}
\arguments{
\item{input}{path to .json configuration file}

\item{output}{path to .Rda/.Rdata}
}
\value{
list
}
\description{
Run from JSON
}
