% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{timeunit.begin.sequence}
\alias{timeunit.begin.sequence}
\title{Sequence of quarter/ month start dates}
\usage{
timeunit.begin.sequence(from, to, by = c("q", "m"))
}
\arguments{
\item{from}{Date}

\item{to}{Date}

\item{by}{"q" or "m"}
}
\value{
Date sequence
}
\description{
Sequence of quarter/ month start dates
}
\examples{
timeunit.begin.sequence(as.Date('2020-01-19'), as.Date('2021-02-07'), by = "q")
# "2020-01-01" "2020-04-01" "2020-07-01" "2020-10-01" "2021-01-01"
}
\keyword{internal}
